+++ akashic-storage/CHANGELOG.md
@@ -1,5 +1,26 @@
 # CHANGELOG
 
+## 1.1.5
+
+- Improve
+  - Valkey Server への再接続設定により接続安定性を向上
+
+## 1.1.4
+
+- Fix
+  - Event 書き込み遅延により、イベントがないものとして応答してしまう問題を修正
+
+## 1.1.3
+
+- Fix
+  - Event 情報が異なるノードに振り分けられてしまう問題を修正
+
+## 1.1.2
+
+- Improve
+  - 部屋を削除するとき、 Valkey Server へ負荷やブロックをかけないクエリに変更
+  - 環境変数 `REQUEST_TIMEOUT` で `requestTimeout` を変更できるように修正
+
 ## 1.1.1
 
 - Bug Fix